WIP localizations for a few more components, cleaned up number format.
