+++ src/components/presentational/ReportContent/ReportContent.tsx
@@ -2,6 +2,7 @@ import { faDownload } from "@fortawesome/free-solid-svg-icons";
 import React from "react";
 import { FontAwesomeSvgIcon } from "react-fontawesome-svg-icon";
 import "./ReportContent.css"
+import { language } from "../../../helpers/language"
 
 export interface ReportContentProps {
     content: string;
@@ -19,7 +20,7 @@ export default function (props: ReportContentProps) {
                 <div className="mdhui-report-content-download-pdf" ng-if="$ctrl.report.ContentType == 'application/pdf'" ng-show="$ctrl.srcSet = true">
                     <a href={"data:application/pdf;base64," + props.content} download={props.type + ".pdf"} title="Download PDF Report">
                         <FontAwesomeSvgIcon icon={faDownload} />
-                        Download PDF Report
+                        {language('download-pdf-report')}
                     </a>
                 </div>
             }
