+++ src/components/step/ConnectDeviceAccountStep/ConnectDeviceAccountStep.tsx
@@ -4,7 +4,7 @@ import StepTitle from "../StepTitle";
 import StepText from "../StepText";
 import StepNextButton from "../StepNextButton";
 import { deviceType } from "../../../helpers/deviceType";
-
+import { language } from "../../../helpers/language";
 export interface ConnectDeviceAccountStepProps {
     title?: string;
     text?: string;
@@ -15,7 +15,7 @@ export interface ConnectDeviceAccountStepProps {
 }
 
 export default function (props: ConnectDeviceAccountStepProps) {
-    const nextButtonText = `Connect to ${props.deviceType}`;
+    const nextButtonText = language("connect-to-device").replace("@@DEVICE@@", props.deviceType);
 
     return (
         <StepLayout>
