+++ src/helpers/locale.ts
@@ -1,6 +1,7 @@
 import { format, formatRelative, parseISO } from 'date-fns';
 import { es, enUS, nl, de, fr, pt, it, pl } from 'date-fns/locale';
 import MyDataHelps from '@careevolution/mydatahelps-js';
+import language from "./language";
 
 export function getLocaleFromIso(language: string): Locale {
     if (language.length < 2) return enUS;
@@ -39,6 +40,19 @@ export function formatRelativeDateForLocale(dateOrDateString: string | Date, bas
     return formatRelative(date, baseDate, { locale: getLocaleFromIso(MyDataHelps.getCurrentLanguage()) });
 }
 
+export function formatNumberForLocale(value: number) {
+    return Number(value.toFixed(0)).toLocaleString(MyDataHelps.getCurrentLanguage())
+}
+
+export function formatMinutesForLocale(value: number) {
+    var hours = Math.floor(value / 60);
+    var displayValue = hours > 0 ? (`${hours}${language("hours-abbreviation")} `) : "";
+    if (Math.round(value - (hours * 60)) !== 0) {
+        displayValue = `${displayValue}${(Math.round(value - (hours * 60)))}${language("minutes-abbreviation")}`;
+    }
+    return displayValue;
+}
+
 export function capitalizeForLocale(str: string) {
     // This won't be adequate if we expand to RTL or symbol-based (e.g., Chinese/Japanese/Korean) languages.
     if (!str) {
@@ -49,7 +63,5 @@ export function capitalizeForLocale(str: string) {
     }
     const firstLetter = str.slice(0, 1).toLocaleUpperCase();
     const rest = str.slice(1);
-    console.log(firstLetter);
-    console.log(rest);
     return `${firstLetter}${rest}`;
 }
